rename modules with Aq prefix and extract common out.
