+++ acquisition/aqClientKit/src/main/java/edu/uci/ics/perpetual/acquisition/datatypes/AcquisitionFunction.java
