+++ acquisition/aqClientKit/src/main/java/edu/uci/ics/perpetual/acquisition/datatypes/RawType.java
@@ -1,0 +1,22 @@
+package edu.uci.ics.perpetual.acquisition.datatypes;
+
+public class RawType {
+    int id;
+    String name;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
