+++ acquisition/aqCommon/src/main/java/edu/uci/ics/perpetual/acquisition/utils/JavaUtils.java
