+++ acquisition/aqCore/src/main/java/edu/uci/ics/perpetual/acquisition/kafka/ObjectSerializer.java
