+++ acquisition/aqCore/src/main/java/edu/uci/ics/perpetual/acquisition/requestmanagement/RequestScheduler.java
@@ -1,0 +1,33 @@
+package edu.uci.ics.perpetual.acquisition.requestmanagement;
+
+import edu.uci.ics.perpetual.acquisition.datatypes.DataSource;
+import edu.uci.ics.perpetual.acquisition.utils.ScheduledStopTask;
+import edu.uci.ics.perpetual.acquisition.datatypes.Request;
+import edu.uci.ics.perpetual.acquisition.datatypes.RequestStatus;
+import edu.uci.ics.perpetual.IngestionThread;
+
+import java.util.Date;
+import java.util.Timer;
+
+public class RequestScheduler {
+
+    public static void scheduleRequest(Request request) throws Exception{
+        // TODO if Policy.isAcquisitionAllowed(request)
+        for(DataSource source : request.getDataSources()){
+            ProducerTask task = new ProducerTask(request,source);
+            ScheduledStopTask stoppingTask = new ScheduledStopTask(task, request.getReqId()+"-stop-task");
+            Timer timer = new Timer("Timer");
+            // TODO LOG
+            System.out.println("Scheduling request: " + request.getReqId() +  " datasource: "+ source.getDsInstanceName() + " datasourceType: "+ source.getDsName());
+            long currentTime = new Date().getTime();
+            timer.schedule(task, request.getStartTime() - currentTime) ;
+            timer.schedule(stoppingTask, request.getEndTime() - currentTime);
+        }
+        // Inform Ingestion Engine scheduled
+        IngestionThread ingestion = new IngestionThread(request.getReqId(),request.getResolution());
+        ingestion.run();
+        request.setStatus(RequestStatus.SCHEDULED);
+        // TODO LOG
+    }
+
+}
