+++ acquisition/aqCore/src/main/resources/acquisition.properties
