--- acquisition/core/src/main/java/edu/uci/ics/perpetual/acquisition/utils/ScheduledStopTask.java
@@ -1,18 +1,0 @@
-package edu.uci.ics.perpetual.acquisition.utils;
-
-import java.util.TimerTask;
-
-public class ScheduledStopTask extends TimerTask {
-
-
-    TimerTask task;
-
-    public ScheduledStopTask(TimerTask task){
-        this.task = task;
-    }
-
-    @Override
-    public void run() {
-        task.cancel();
-    }
-}
