+++ acquisition/pom.xml
@@ -10,8 +10,9 @@
     <modelVersion>4.0.0</modelVersion>
     <packaging>pom</packaging>
 <modules>
-    <module>clientKit</module>
-    <module>core</module>
+    <module>aqCommon</module>
+    <module>aqClientKit</module>
+    <module>aqCore</module>
 </modules>
     <artifactId>acquisition</artifactId>
     <build>
